# Speed up curve plotting