Fixes #61.

- broke up `update_plot()` method in data_view_box.py by creating `_update_traces()` and `_update_colormap()` methods
- changed the way curves are plotted: `PlotCurveItem` objects are generated for number of channels requested, and the data held in each object is subsequently updated with every call to `_update_traces()` resulting in much faster plotting
- disabled auto range for the plot - `autoRange()` is now performed once at the end after plotting
- major speed up in `get_predicted_traces()`
  - used slightly faster variants of existing numpy commands
  - moved cupy calls outside function to make the function numba compatible
  - used numba to drastically speed up on-the-fly operation (results in a small delay the first time function is called)

A few quick benchmarks of the `update_plot()` method for a before (master branch) and after (current PR) comparison:

<table align="center">
<thead>
  <tr>
    <th rowspan="2">selected views</th>
    <th colspan="2">traces</th>
    <th colspan="2">colormap</th>
  </tr>
  <tr>
    <td>before</td>
    <td>after</td>
    <td>before</td>
    <td>after</td>
  </tr>
</thead>
<tbody>
  <tr>
    <td>raw</td>
    <td>200ms</td>
    <td>23ms</td>
    <td>4ms</td>
    <td>16ms</td>
  </tr>
  <tr>
    <td>whitened</td>
    <td>207ms</td>
    <td>30ms</td>
    <td>16ms</td>
    <td>40ms</td>
  </tr>
  <tr>
    <td>prediction</td>
    <td>216ms</td>
    <td>48ms</td>
    <td>48ms</td>
    <td>47ms</td>
  </tr>
  <tr>
    <td>residual</td>
    <td>229ms</td>
    <td>55ms</td>
    <td>43ms</td>
    <td>53ms</td>
  </tr>
  <tr>
    <td>all</td>
    <td>1.35s</td>
    <td>68ms</td>
    <td>-</td>
    <td>-</td>
  </tr>
</tbody>
</table>

There is significant improvement in curve plotting speed while sacrificing some speed in colormap plotting.

Additionally, the curve plotting speed is not adversely affected by the number of channels being plotted:

<table align="center">
<thead>
  <tr>
    <th rowspan="2">number<br>of<br>traces<br>(raw)</th>
    <th colspan="2">time</th>
  </tr>
  <tr>
    <td>before<br></td>
    <td>after</td>
  </tr>
</thead>
<tbody>
  <tr>
    <td>32</td>
    <td>158ms</td>
    <td>22ms</td>
  </tr>
  <tr>
    <td>62</td>
    <td>557ms</td>
    <td>31ms</td>
  </tr>
  <tr>
    <td>122</td>
    <td>2073ms</td>
    <td>50ms</td>
  </tr>
  <tr>
    <td>242</td>
    <td>7917ms</td>
    <td>80ms</td>
  </tr>
</tbody>
</table>